Changes the forced subtitle selection to a more appropiate behavior
Kodi selects forced subtitles based on the audio stream that was selected according to a user's configuration. If no audio language exist inside a container matching the users configuration no subtitle is automatically selected even if a forced subtitle for the users language is present in a cotainer. This patch addresses this and Kodi will do a second check for a subtitle that has the default and forced flag set and matching the user's selected language. If a match is found it will select it.

The check for the default flag was added for the 2nd check as additional means for users to control the behavior in case they understand more than one language. A foced track of a second language can still be present and properly marked as forced without being displayed when no matching audio track in present.